Moved all schemas to models/schemas.js to reduce clutter in routes.js
